add defl.c and def.c
